requirements.txt
easyocr==1.3.0.1
imageio==2.9.0
numpy==1.20.2
opencv-python==4.5.1.48
streamlit==0.80.0
streamlit-webcam-example==0.1.0
torch==1.8.1
torchvision==0.9.1
logger==1.4